when source.update() returns False to indicate an error, handle that like an exception rather than storing the source update time as if it worked
git-svn-id: https://svn.ritchielab.psu.edu/svn/projects/LOKI@3059 9e62ecf6-770d-0410-a9d7-e62a33e3d092

+++ branches/2.x-omicunits/loki/loki_updater.py
@@ -248,7 +248,8 @@ class Updater(object):
 						self.logPush("processing %s data ...\n" % srcName)
 						
 						cursor.execute("DELETE FROM `db`.`warning` WHERE source_id = ?", (srcID,))
-						srcObj.update(options)
+						if srcObj.update(options) == False:
+							raise Exception
 						cursor.execute("UPDATE `db`.`source` SET updated = DATETIME('now'), version = ? WHERE source_id = ?", (srcObj.getVersionString(), srcID))
 						
 						cursor.execute("DELETE FROM `db`.`source_option` WHERE source_id = ?", (srcID,))
